Move more methods to NumericHelper
