+++ src/NumericHelper.php
@@ -1,7 +1,10 @@
 <?php
-
 namespace Yiisoft\Strings;
 
+
+/**
+ * Provides static methods to work with numeric strings.
+ */
 final class NumericHelper
 {
     /**
@@ -24,5 +27,39 @@ final class NumericHelper
             default:
                 return $number . 'th';
         }
+    }
+
+    /**
+     * Returns string representation of number value with replaced commas to dots, if decimal point
+     * of current locale is comma.
+     * @param int|float|string $value
+     * @return string
+     */
+    public static function normalizeNumber($value): string
+    {
+        $value = (string)$value;
+
+        $localeInfo = localeconv();
+        $decimalSeparator = $localeInfo['decimal_point'] ?? null;
+
+        if ($decimalSeparator !== null && $decimalSeparator !== '.') {
+            $value = str_replace($decimalSeparator, '.', $value);
+        }
+
+        return $value;
+    }
+
+    /**
+     * Safely casts a float to string independent of the current locale.
+     *
+     * The decimal separator will always be `.`.
+     * @param float|int $number A floating point number or integer.
+     * @return string The string representation of the number.
+     */
+    public static function floatToString($number): string
+    {
+        // . and , are the only decimal separators known in ICU data,
+        // so its safe to call str_replace here
+        return str_replace(',', '.', (string)$number);
     }
 }
