+++ src/StringHelper.php
@@ -5,14 +5,13 @@ declare(strict_types=1);
 namespace Yiisoft\Strings;
 
 use function array_slice;
-use function htmlspecialchars;
 use function mb_strlen;
 use function mb_strtolower;
 use function mb_strtoupper;
 use function mb_substr;
 
 /**
- * Provides static methods allowing you to deal with strings more efficiently.
+ * Provides static methods to work with strings.
  */
 final class StringHelper
 {
@@ -397,40 +396,6 @@ final class StringHelper
         }
 
         return $result;
-    }
-
-    /**
-     * Returns string representation of number value with replaced commas to dots, if decimal point
-     * of current locale is comma.
-     * @param int|float|string $value
-     * @return string
-     */
-    public static function normalizeNumber($value): string
-    {
-        $value = (string)$value;
-
-        $localeInfo = localeconv();
-        $decimalSeparator = $localeInfo['decimal_point'] ?? null;
-
-        if ($decimalSeparator !== null && $decimalSeparator !== '.') {
-            $value = str_replace($decimalSeparator, '.', $value);
-        }
-
-        return $value;
-    }
-
-    /**
-     * Safely casts a float to string independent of the current locale.
-     *
-     * The decimal separator will always be `.`.
-     * @param float|int $number A floating point number or integer.
-     * @return string The string representation of the number.
-     */
-    public static function floatToString($number): string
-    {
-        // . and , are the only decimal separators known in ICU data,
-        // so its safe to call str_replace here
-        return str_replace(',', '.', (string) $number);
     }
 
     /**
