+++ tests/NumericHelperTest.php
@@ -17,4 +17,20 @@ final class NumericHelperTest extends TestCase
         $this->assertEquals('111th', NumericHelper::toOrdinal(111));
         $this->assertEquals('113th', NumericHelper::toOrdinal(113));
     }
+
+    public function testFloatToString(): void
+    {
+        $this->assertSame('10.111', NumericHelper::floatToString('10,111'));
+    }
+
+    public function testNormalizeNumber(): void
+    {
+        $setLocale = setlocale(LC_ALL, 'Norwegian');
+
+        if (!$setLocale) {
+            $this->markTestSkipped('Norwegian locale not found.');
+        }
+
+        $this->assertSame('10.000', NumericHelper::normalizeNumber('10,000'));
+    }
 }
